Move featuresToSet out of shared functions
This is to work around an issue with 3rd party libraries in shared code,
but this is also a better spot for this function since it's only used by
selection tools.
